+++ CHANGELOG.md
@@ -12,8 +12,17 @@
 
 * See liked / disliked / favorited posts
 * Like / dislike / Favorited posts
-
 * Sped up page loads, removed connected animation
 * Pools search added
 * Pool view added
 * Pools button now shows and links to pools that the post is in
+
+# Version 0.6
+
+* Added more pool info in search
+* Added recommended tags button
+* Fixed various bugs
+* Added Comments to posts
+* You can now comment on posts
+* Some dtext is converted to markdown for proper display
+* Moved post controls into overflow menu